Compatibility changes for VaaS and Dataverse (#35)
* Use filename as identifier

* Revert some file logic, moved to tasks instead of service

Added an integration test for the CollectTask

* Add more logging

* Fix recursive ocfl tar task creation

Filter out anything that is not a file on the root of the inbox, for any
inbox. This caused some issues with ocfl tar tasks being generated and a
huge tree of rejected/rejected/rejected etc.

Also verified that an inbox item is actually known in the database,
otherwise rejecting it.

* Add new metadata reader based on Jena

* Apply feedback, rename transfer item fields

* Remove bagDepositDate as queueDate was already there

+++ src/main/java/nl/knaw/dans/ttv/core/service/FileServiceImpl.java
@@ -17,13 +17,16 @@ package nl.knaw.dans.ttv.core.service;
 
 import org.apache.commons.codec.digest.DigestUtils;
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.FilenameUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.FileNotFoundException;
+import java.io.BufferedInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.PrintWriter;
+import java.io.StringWriter;
+import java.nio.file.FileAlreadyExistsException;
 import java.nio.file.FileStore;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -38,6 +41,11 @@ public class FileServiceImpl implements FileService {
         Objects.requireNonNull(current, "current path cannot be null");
         Objects.requireNonNull(newPath, "newPath cannot be null");
         log.trace("moving file from {} to {}", current, newPath);
+
+        if (Files.exists(newPath)) {
+            throw new FileAlreadyExistsException(String.format("Cannot move file %s to %s, file already exists", current, newPath));
+        }
+
         return Files.move(current, newPath);
     }
 
@@ -46,11 +54,20 @@ public class FileServiceImpl implements FileService {
         Objects.requireNonNull(filePath, "filePath cannot be null");
         Objects.requireNonNull(newPath, "newPath cannot be null");
 
-        var tempTarget = Path.of(newPath + ".part");
+        var tempTarget = newPath.getParent().resolve(newPath.getFileName() + ".part");
+
+        var store1 = Files.getFileStore(filePath);
+        var store2 = Files.getFileStore(newPath.getParent());
+
+        if (store1.equals(store2)) {
+            log.info("Moving file {} to {}", filePath, newPath);
+            return moveFile(filePath, newPath);
+        }
 
         // there could be leftovers from a previous attempt, remove them
         Files.deleteIfExists(tempTarget);
 
+        log.info("Moving file atomically {} to {}", filePath, newPath);
         moveFile(filePath, tempTarget);
         return moveFile(tempTarget, newPath);
     }
@@ -76,30 +93,32 @@ public class FileServiceImpl implements FileService {
         var rejectedFolder = path.getParent().resolve("rejected");
         ensureDirectoryExists(rejectedFolder);
 
-        log.debug("Moving rejected file to '{}'", rejectedFolder);
-        var index = path.getFileName().toString().lastIndexOf(".");
-        var extension = path.getFileName().toString().substring(index);
-        var fileBaseName = path.getFileName().toString().substring(0, index);
+        var filename = path.getFileName().toString();
 
-        log.trace("File base name is '{}', extension is '{}'", extension, fileBaseName);
-        var rejectedFileName = fileBaseName + extension;
+        log.debug("Moving rejected file to '{}'", rejectedFolder);
+        var extension = FilenameUtils.getExtension(filename);
+        var fileBaseName = FilenameUtils.removeExtension(filename);
+
+        log.trace("File base name is '{}', extension is '{}'", fileBaseName, extension);
+        var rejectedFileName = fileBaseName + "." + extension;
         var errorFileName = fileBaseName + ".error.txt";
 
         var duplicateCounter = 1;
 
-        while (Files.exists(Path.of(rejectedFolder.toString(), rejectedFileName))) {
-            log.trace("File '{}' already exists, generating a new filename", Path.of(rejectedFolder.toString(), rejectedFileName));
-            rejectedFileName = fileBaseName + "_" + duplicateCounter + extension;
+        while (Files.exists(rejectedFolder.resolve(rejectedFileName))) {
+            log.trace("File '{}' already exists, generating a new filename", rejectedFolder.resolve(rejectedFileName));
+            rejectedFileName = fileBaseName + "_" + duplicateCounter + "." + extension;
             errorFileName = fileBaseName + "_" + duplicateCounter + ".error.txt";
 
             duplicateCounter += 1;
         }
 
         log.trace("Settled on '{}' for filename", rejectedFileName);
-        var targetPath = Path.of(rejectedFolder.toString(), rejectedFileName);
-        var targetErrorPath = Path.of(rejectedFolder.toString(), errorFileName);
+        var targetPath = rejectedFolder.resolve(rejectedFileName);
+        var targetErrorPath = rejectedFolder.resolve(errorFileName);
 
         log.trace("Moving file to '{}', writing error report to '{}'", targetPath, targetErrorPath);
+
         Files.move(path, targetPath);
         writeExceptionToFile(targetErrorPath, exception);
     }
@@ -124,10 +143,14 @@ public class FileServiceImpl implements FileService {
         return Files.getFileStore(path);
     }
 
-    void writeExceptionToFile(Path errorReportName, Exception exception) throws FileNotFoundException {
-        var writer = new PrintWriter(errorReportName.toFile());
-        exception.printStackTrace(writer);
-        writer.close();
+    void writeExceptionToFile(Path errorReportName, Exception exception) throws IOException {
+        var writer = new StringWriter();
+        var pw = new PrintWriter(writer);
+        exception.printStackTrace(pw);
+        var stackTrace = writer.toString();
+        pw.close();
+
+        Files.write(errorReportName, stackTrace.getBytes());
     }
 
     @Override
@@ -156,7 +179,10 @@ public class FileServiceImpl implements FileService {
     public String calculateChecksum(Path path) throws IOException {
         Objects.requireNonNull(path, "path cannot be null");
         log.trace("Calculating checksum for '{}'", path);
-        return new DigestUtils("SHA-256").digestAsHex(Files.readAllBytes(path));
+
+        try (var input = Files.newInputStream(path)) {
+            return DigestUtils.sha256Hex(new BufferedInputStream(input));
+        }
     }
 
     @Override
@@ -169,16 +195,22 @@ public class FileServiceImpl implements FileService {
     @Override
     public long getPathSize(Path path) throws IOException {
         Objects.requireNonNull(path, "path cannot be null");
-        return Files.walk(path).filter(Files::isRegularFile).map(p -> {
-            try {
-                var size = getFileSize(p);
-                log.trace("File size for file '{}' is {} bytes", p, size);
-                return size;
-            }
-            catch (IOException e) {
-                throw new RuntimeException(e);
-            }
-        }).reduce(0L, Long::sum);
+
+        try (var files = Files.walk(path)) {
+            return files
+                .filter(Files::isRegularFile)
+                .mapToLong(p -> {
+                    try {
+                        var size = getFileSize(p);
+                        log.trace("File size for file '{}' is {} bytes", p, size);
+                        return size;
+                    }
+                    catch (IOException e) {
+                        throw new RuntimeException(e);
+                    }
+                })
+                .sum();
+        }
     }
 
     @Override
